+++ tests/Python/TestTimeSpan.fs
@@ -233,3 +233,442 @@ let ``test TimeSpan implementation coherence`` () =
 //    test_float 2200. 11. 200.
 //    test_float -3000. 1.5 -2000.
 //    test_float 0. 1000. 0.
+
+
+    // **************************************************
+    // Test Cases From :
+    // https://docs.microsoft.com/en-us/dotnet/api/system.timespan.tryparse
+    //
+    // **************************************************
+    //
+    //            String to Parse                TimeSpan
+    //            ---------------   ---------------------
+    //                          0        00:00:00
+    //                         14     14.00:00:00
+    //                      1:2:3        01:02:03
+    //                  0:0:0.250        00:00:00.2500000
+    //            10.20:30:40.050     10.20:30:40.0050000
+    //        99.23:59:59.9990000     99.23:59:59.9999000
+    //        0023:0059:0059.0009        23:59:59.0009000
+    //                     23:0:0        23:00:00
+    //                     24:0:0   Parse operation failed. (actually not true, it's parsed to 24 days...)
+    //                     0:59:0        00:59:00
+    //                     0:60:0   Parse operation failed.
+    //                     0:0:59        00:00:59
+    //                     0:0:60   Parse operation failed.
+    //                        10:   Parse operation failed.
+    //                       10:0        10:00:00
+    //                        :10   Parse operation failed.
+    //                       0:10        00:10:00
+    //                     10:20:   Parse operation failed.
+    //                    10:20:0        10:20:00
+    //                       .123   Parse operation failed.
+    //                    0.12:00        12:00:00
+    //                        10.   Parse operation failed.
+    //                      10.12   Parse operation failed.
+    //                   10.12:00     10.12:00:00
+
+[<Fact>]
+let ``test TimeSpan 0 parse works`` () =
+    let actual = TimeSpan.Parse("0")
+    let expected = TimeSpan(0, 0, 0, 0, 0)
+    equal actual expected
+
+[<Fact>]
+let ``test TimeSpan 14 parse works`` () =
+    let actual = TimeSpan.Parse("14")
+    let expected = TimeSpan(14, 0, 0, 0, 0)
+    equal actual expected
+
+[<Fact>]
+let ``test TimeSpan 1:2:3 parse works`` () =
+    let actual = TimeSpan.Parse("1:2:3", CultureInfo.InvariantCulture)
+    let expected = TimeSpan(0, 1, 2, 3, 0)
+    equal actual expected
+
+[<Fact>]
+let ``test TimeSpan 0:0:0.250 parse works`` () =
+    let actual = TimeSpan.Parse("0:0:0.250")
+    let expected = TimeSpan(0, 0, 0, 0, 250)
+    equal actual expected
+
+[<Fact>]
+let ``test TimeSpan 10.20:30:40.050 parse works`` () =
+    let actual = TimeSpan.Parse("10.20:30:40.050")
+    let expected = TimeSpan(10, 20, 30, 40, 50)
+    equal actual expected
+
+[<Fact>]
+let ``test TimeSpan 99.23:59:59.999 parse works`` () =
+    let actual = TimeSpan.Parse("99.23:59:59.999")
+    let expected = TimeSpan(99, 23, 59, 59, 999)
+    equal actual expected
+
+[<Fact>]
+let ``test TimeSpan 0023:0059:0059.0099 parse works`` () =
+    let actual = TimeSpan.Parse("0023:0059:0059.009")
+    let expected = TimeSpan(00, 23, 59, 59, 9)
+    equal actual expected
+
+[<Fact>]
+let ``test TimeSpan 23:0:0 parse works`` () =
+    let actual = TimeSpan.Parse("23:0:0")
+    let expected = TimeSpan(0, 23, 0, 0, 0)
+    equal actual expected
+
+#if FABLE_COMPILER
+// msft assumes it's 24 days...
+// https://docs.microsoft.com/en-us/dotnet/api/system.timespan.parse
+// or
+// https://github.com/dotnet/dotnet-api-docs/blob/7f6a3882631bc008b858adfadb43cd17bbd55d49/xml/System/TimeSpan.xml#L2772
+[<Fact>]
+let ``test TimeSpan 24:0:0 parse fails`` () =
+    (fun _ -> TimeSpan.Parse("24:0:0"))
+    |> Util.throwsError "String '24:0:0' was not recognized as a valid TimeSpan."
+
+[<Fact>]
+let ``test TimeSpan 24:0:0 TryParse fails`` () =
+    let status, _ = TimeSpan.TryParse("24:0:0", CultureInfo.InvariantCulture)
+    equal status false
+#endif
+
+[<Fact>]
+let ``test TimeSpan 0:0:59 parse works`` () =
+    let actual = TimeSpan.Parse("0:0:59")
+    let expected = TimeSpan(0, 0, 0, 59, 0)
+    equal actual expected
+
+#if FABLE_COMPILER // temporary, TODO: fix test to be backwards compatible
+[<Fact>]
+let ``test TimeSpan 0:60:0 parse fails`` () =
+    (fun _ -> TimeSpan.Parse("0:60:0"))
+#if FABLE_COMPILER
+    |> Util.throwsError "String '0:60:0' was not recognized as a valid TimeSpan."
+#else
+    |> Util.throwsError "The TimeSpan string '0:60:0' could not be parsed because at least one of the numeric components is out of range or contains too many digits."
+#endif
+#endif
+
+#if FABLE_COMPILER // temporary, TODO: fix test to be backwards compatible
+[<Fact>]
+let ``test TimeSpan 10: parse fails`` () =
+    (fun _ -> TimeSpan.Parse("10:"))
+    |> Util.throwsError "String '10:' was not recognized as a valid TimeSpan."
+#endif
+
+[<Fact>]
+let ``test TimeSpan 10:0 parse works`` () =
+    let actual = TimeSpan.Parse("10:0")
+    let expected = TimeSpan(0, 10, 0, 0, 0)
+    equal actual expected
+
+#if FABLE_COMPILER // temporary, TODO: fix test to be backwards compatible
+[<Fact>]
+let ``test TimeSpan 10:20: parse fails`` () =
+    (fun _ -> TimeSpan.Parse("10:20:"))
+    |> Util.throwsError "String '10:20:' was not recognized as a valid TimeSpan."
+#endif
+
+[<Fact>]
+let ``test TimeSpan 10:20:0 parse works`` () =
+    let actual = TimeSpan.Parse("10:20:0")
+    let expected = TimeSpan(0, 10, 20, 0, 0)
+    equal actual expected
+
+
+#if FABLE_COMPILER // temporary, TODO: fix test to be backwards compatible
+[<Fact>]
+let ``test TimeSpan .123 parse fails`` () =
+    (fun _ -> TimeSpan.Parse(".123"))
+    |> Util.throwsError "String '.123' was not recognized as a valid TimeSpan."
+#endif
+
+[<Fact>]
+let ``test TimeSpan 0.12:00 parse works`` () =
+    let actual = TimeSpan.Parse("0.12:00")
+    let expected = TimeSpan(0, 12, 00, 0, 0)
+    equal actual expected
+
+#if FABLE_COMPILER // temporary, TODO: fix test to be backwards compatible
+[<Fact>]
+let ``test TimeSpan 10. parse fails`` () =
+    (fun _ -> TimeSpan.Parse("10."))
+    |> Util.throwsError "String '10.' was not recognized as a valid TimeSpan."
+#endif
+
+#if FABLE_COMPILER // temporary, TODO: fix test to be backwards compatible
+[<Fact>]
+let ``test TimeSpan 10.12 parse fails`` () =
+    (fun _ -> TimeSpan.Parse("10.12"))
+    |> Util.throwsError "String '10.12' was not recognized as a valid TimeSpan."
+#endif
+
+[<Fact>]
+let ``test TimeSpan 10.12:00 parse works`` () =
+    let actual = TimeSpan.Parse("10.12:00")
+    let expected = TimeSpan(10, 12, 00, 0, 0)
+    equal actual expected
+
+[<Fact>]
+let ``test TimeSpan 0 TryParse works`` () =
+    let status, actual = TimeSpan.TryParse("0")
+    let expected = TimeSpan(0, 0, 0, 0, 0)
+    equal status true
+    equal actual expected
+
+[<Fact>]
+let ``test TimeSpan 14 TryParse works`` () =
+    let status, actual = TimeSpan.TryParse("14")
+    let expected = TimeSpan(14, 0, 0, 0, 0)
+    equal status true
+    equal actual expected
+
+[<Fact>]
+let ``test TimeSpan 1:2:3 TryParse works`` () =
+    let status, actual = TimeSpan.TryParse("1:2:3")
+    let expected = TimeSpan(0, 1, 2, 3, 0)
+    equal status true
+    equal actual expected
+
+[<Fact>]
+let ``test TimeSpan 0:0:0.250 TryParse works`` () =
+    let status, actual = TimeSpan.TryParse("0:0:0.250")
+    let expected = TimeSpan(0, 0, 0, 0, 250)
+    equal status true
+    equal actual expected
+
+[<Fact>]
+let ``test TimeSpan 10.20:30:40.050 TryParse works`` () =
+    let status, actual = TimeSpan.TryParse("10.20:30:40.050")
+    let expected = TimeSpan(10, 20, 30, 40, 50)
+    equal status true
+    equal actual expected
+
+[<Fact>]
+let ``test TimeSpan 99.23:59:59.999 TryParse works`` () =
+    let status, actual = TimeSpan.TryParse("99.23:59:59.999")
+    let expected = TimeSpan(99, 23, 59, 59, 999)
+    equal status true
+    equal actual expected
+
+[<Fact>]
+let ``test TimeSpan 0023:0059:0059.009 TryParse works`` () =
+    let status, actual = TimeSpan.TryParse("0023:0059:0059.009")
+    let expected = TimeSpan(00, 23, 59, 59, 9)
+    equal status true
+    equal actual expected
+
+[<Fact>]
+let ``test TimeSpan 23:0:0 TryParse works`` () =
+    let status, actual = TimeSpan.TryParse("23:0:0")
+    let expected = TimeSpan(0, 23, 0, 0, 0)
+    equal status true
+    equal actual expected
+
+[<Fact>]
+let ``test TimeSpan 0:0:59 TryParse works`` () =
+    let status, actual = TimeSpan.TryParse("0:0:59")
+    let expected = TimeSpan(0, 0, 0, 59, 0)
+    equal status true
+    equal actual expected
+
+[<Fact>]
+let ``test TimeSpan 0:60:0 TryParse fails`` () =
+    let status, _ = TimeSpan.TryParse("0:60:0")
+    equal status false
+
+[<Fact>]
+let ``test TimeSpan 10: TryParse fails`` () =
+    let status, _ = TimeSpan.TryParse("10:")
+    equal status false
+[<Fact>]
+let ``test TimeSpan 10:0 TryParse works`` () =
+    let status, actual = TimeSpan.TryParse("10:0")
+    let expected = TimeSpan(0, 10, 0, 0, 0)
+    equal status true
+    equal actual expected
+
+[<Fact>]
+let ``test TimeSpan 10:20: TryParse fails`` () =
+    let status, _ = TimeSpan.TryParse("10:20:")
+    equal status false
+
+[<Fact>]
+let ``test TimeSpan 10:20:0 TryParse works`` () =
+    let status, actual = TimeSpan.TryParse("10:20:0")
+    let expected = TimeSpan(0, 10, 20, 0, 0)
+    equal status true
+    equal actual expected
+
+[<Fact>]
+let ``test TimeSpan .123 TryParse fails`` () =
+    let status, _ = TimeSpan.TryParse(".123")
+    equal status false
+
+[<Fact>]
+let ``test TimeSpan 0.12:00 TryParse works`` () =
+    let status, actual = TimeSpan.TryParse("0.12:00")
+    let expected = TimeSpan(0, 12, 00, 0, 0)
+    equal status true
+    equal actual expected
+
+[<Fact>]
+let ``test TimeSpan 10. TryParse fails`` () =
+    let status, _ = TimeSpan.TryParse("10.")
+    equal status false
+
+[<Fact>]
+let ``test TimeSpan 10.12 TryParse fails`` () =
+    let status, _ = TimeSpan.TryParse("10.12")
+    equal status false
+
+[<Fact>]
+let ``test TimeSpan 10.12:00 TryParse works`` () =
+    let status, actual = TimeSpan.TryParse("10.12:00")
+    let expected = TimeSpan(10, 12, 00, 0, 0)
+    equal status true
+    equal actual expected
+
+[<Fact>]
+let ``test TimeSpan 00:00:00.1 Parse handle correctly the milliseconds`` () =
+    let actual = TimeSpan.Parse("00:00:00.1").TotalMilliseconds
+    let expected = 100.
+    equal actual expected
+
+[<Fact>]
+let ``test TimeSpan 00:00:00.12 Parse handle correctly the milliseconds`` () =
+    let actual = TimeSpan.Parse("00:00:00.12").TotalMilliseconds
+    let expected = 120.
+    equal actual expected
+
+[<Fact>]
+let ``test TimeSpan 00:00:00.123 Parse handle correctly the milliseconds`` () =
+    let actual = TimeSpan.Parse("00:00:00.123").TotalMilliseconds
+    let expected = 123.
+    equal actual expected
+
+[<Fact>]
+let ``test TimeSpan 00:00:00.1234 Parse handle correctly the milliseconds`` () =
+    let actual = TimeSpan.Parse("00:00:00.1234").TotalMilliseconds
+    let expected = 123.4
+    equal actual expected
+
+[<Fact>]
+let ``test TimeSpan 00:00:00.12345 Parse handle correctly the milliseconds`` () =
+    let actual = TimeSpan.Parse("00:00:00.12345").TotalMilliseconds
+    let expected = 123.45
+    equal actual expected
+
+[<Fact>]
+let ``test TimeSpan 00:00:00.123456 Parse handle correctly the milliseconds`` () =
+    let actual = TimeSpan.Parse("00:00:00.123456").TotalMilliseconds
+    let expected = 123.456
+    equal actual expected
+
+[<Fact>]
+let ``test TimeSpan 00:00:00.0034567 Parse handle correctly the milliseconds`` () =
+    let actual = TimeSpan.Parse("00:00:00.0034567").TotalMilliseconds
+    let expected = 3.4567
+    equal actual expected
+
+[<Fact>]
+let ``test TimeSpan Parse work with negative TimeSpan`` () =
+    let actual = TimeSpan.Parse("-2:00:00")
+    equal actual.TotalMilliseconds -7200000.0
+
+[<Fact>]
+let ``test TimeSpan 1.23:45:06.789 TotalMilliseconds works`` () =
+    let actual = TimeSpan.Parse("1.23:45:06.789").TotalMilliseconds
+    let expected = 171906789.0
+    equal actual expected
+
+[<Fact>]
+let ``test TimeSpan 1.23:45:06.789 TotalSeconds works`` () =
+    let actual = TimeSpan.Parse("1.23:45:06.789").TotalSeconds
+    let expected = 171906.789
+    equal actual expected
+
+[<Fact>]
+let ``test TimeSpan 1.23:45:06.789 TotalMinutes works`` () =
+    let actual = TimeSpan.Parse("1.23:45:06.789").TotalMinutes
+    let expected = 2865.11315
+    equal actual expected
+
+[<Fact>]
+let ``test TimeSpan 1.23:45:06.789 TotalHours works`` () =
+    let actual = TimeSpan.Parse("1.23:45:06.789").TotalHours
+    let expected = 47.75188583333333
+    equal actual expected
+
+[<Fact>]
+let ``test TimeSpan 1.23:45:06.789 TotalDays works`` () =
+    let actual = TimeSpan.Parse("1.23:45:06.789").TotalDays
+    let expected = 1.9896619097222221
+    equal actual expected
+
+[<Fact>]
+let ``test TimeSpan TotalSeconds & friends work`` () =
+    let ts = TimeSpan.FromDays(0.005277777778)
+    ts.TotalMilliseconds |> equal 456000.0
+    ts.TotalSeconds |> equal 456.
+    ts.TotalMinutes |> equal 7.6
+
+[<Fact>]
+let ``test TimeSpan.Milliseconds works with positive TimeSpan`` () =
+    let actual = TimeSpan.Parse("1.23:45:06.789").Milliseconds
+    let expected = 789
+    equal actual expected
+
+[<Fact>]
+let ``test TimeSpan.Milliseconds works with negative TimeSpan`` () =
+    let actual = TimeSpan.Parse("-1.23:45:06.78999").Milliseconds
+    let expected = -789
+    equal actual expected
+
+[<Fact>]
+let ``test TimeSpan.Seconds works with positive TimeSpan`` () =
+    let actual = TimeSpan.Parse("1.23:45:06.789").Seconds
+    let expected = 6
+    equal actual expected
+
+[<Fact>]
+let ``test TimeSpan.Seconds works with negative TimeSpan`` () =
+    let actual = TimeSpan.Parse("-1.23:45:06.78999").Seconds
+    let expected = -6
+    equal actual expected
+
+[<Fact>]
+let ``test TimeSpan.Minutes works with positive TimeSpan`` () =
+    let actual = TimeSpan.Parse("1.23:45:06.789").Minutes
+    let expected = 45
+    equal actual expected
+
+[<Fact>]
+let ``test TimeSpan.Minutes works with negative TimeSpan`` () =
+    let actual = TimeSpan.Parse("-1.23:45:06.78999").Minutes
+    let expected = -45
+    equal actual expected
+
+[<Fact>]
+let ``test TimeSpan.Hours works with positive TimeSpan`` () =
+    let actual = TimeSpan.Parse("1.23:45:06.789").Hours
+    let expected = 23
+    equal actual expected
+
+[<Fact>]
+let ``test TimeSpan.Hours works with negative TimeSpan`` () =
+    let actual = TimeSpan.Parse("-1.23:45:06.78999").Hours
+    let expected = -23
+    equal actual expected
+
+[<Fact>]
+let ``test TimeSpan.Days works with positive TimeSpan`` () =
+    let actual = TimeSpan.Parse("1.23:45:06.789").Days
+    let expected = 1
+    equal actual expected
+
+[<Fact>]
+let ``test TimeSpan.Days works with negative TimeSpan`` () =
+    let actual = TimeSpan.Parse("-1.23:45:06.78999").Days
+    let expected = -1
+    equal actual expected